requirements.txt
# Web framework (for later API phase)
fastapi
uvicorn[standard]
python-multipart
httpx

# Database
sqlalchemy
psycopg2-binary

# Migrations
alembic


# pgvector support in SQLAlchemy
pgvector

# Environment variables
python-dotenv
pyyaml

# Data processing and utilities
pandas
requests
tqdm
pillow
boto3
scipy
sounddevice

# Vector/embedding and local DB tooling
chromadb

# Generative AI / LLM integrations
google-genai
groq
edge-tts

# GPU-Specific Libraries
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio
tensorflow[and-cuda]

# NLP and evaluation tooling (used in notebooks)
numpy
scikit-learn
spacy
textstat
langdetect
sentence-transformers
transformers
torch
joblib

# Jupyter for experimentation
notebook
ipykernel

# Optional but very useful
pydantic
